Enum dazeus::ConfigGroup

pub enum ConfigGroup {
    Plugin,
    Core,
}

The type of config that should be retrieved.

Variants

Plugin

Indicates a config value that should be retrieved from the plugin settings.

Core

Indicates a config value that should be retrieved from the core settings.